backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
python-docx==1.1.2
PyPDF2==3.0.1
chardet==5.2.0
httpx==0.27.2
openai==1.51.0
python-dotenv==1.0.1
